Update color palette
Color palette is updated after issue was revisited on twitter. Thanks to pictures from David Haywood (@mamehaze), thanks @Toryalai1 for pointing me to the twitter thread.

The new palette over HDMI more accurately represents what it looks like on the CRT picture provided. I changed that to the default. The alternate with a more light blue ocean is available as the second option.