add crt images
dsc-7007 and dsc-7005 are from kold666, calibrated_crt_arcade is from mamehaze, showing the difference between likely a calibrated crt and the one kold presented. it also could be different proms, one with a bug, one without. not much is known about this board to this level of detail and it's a very rare board.